requirements.txt
flask
werkzeug
supabase
python-dotenv